Still working on the Board
In the process of trying to solve the problem of representing the State
of the board, other problems surfaced in other classes that made the
Board problem problematic. In order to solve these problems, a lot of
refactoring was done on the classes where changes took place and a few
classes were also added.