AI fully functional.
Due to the limited amount of time available and the amount of work that
is left, I have decided to leave the AI functionality as it is with it's
obvious bugs and all. Note it is highly capable but prone to some
unpleasant bugs. This would be addressed when all other high priority
work is done.